InsiderOUT.PythonService/requirements.txt
fastapi
uvicorn[standard]
python-docx
lxml
pillow
fpdf
pikepdf
google-generativeai